refactor(SessionSidebar): 抽取项目选择器与 worktree 切换器
将主体中两个较大的下拉区域移入纯展示型的 ProjectPicker 与 WorktreeSwitcher
组件。所有 state、ref、effect 和 handler 仍留在 SessionSidebar 中并以 props
传入，因此共享的点击外部关闭 effect 依旧同时持有两个下拉的 ref——行为保持不变。
共享的 WorktreeEntry/WorktreeState 类型移至 session-sidebar/types.ts。

SessionSidebar.tsx：1318 -> 770 行（最初为 2023 行），现已低于 1000 行。
