+++ components/session-sidebar/types.ts
@@ -1,0 +1,16 @@
+export interface WorktreeEntry {
+  path: string;
+  branch: string | null;
+  isMain: boolean;
+}
+
+export interface WorktreeState {
+  /** The cwd this data was fetched for — guards against stale responses */
+  forCwd: string;
+  projectRoot: string;
+  isGit: boolean;
+  /** False when forCwd is a repo subdirectory — the switcher is hidden there
+   *  because subdir sessions keep their own project identity */
+  isTopLevel: boolean;
+  worktrees: WorktreeEntry[];
+}
